improvement: move edge utils to its own file
